Fix handling of non-associated ExtractedSources.
In theory, all ExtractedSources should have an entry in the associated sources table;
since even if the source is a singleton it should be self-associated.

However, if the trap failed part way through so source association was not performed,
_or if the source association routines are buggy_,
then there may be extracted sources without an associated id.

Hence I've implemented a left outer join and handling of null values.