Work around for truncated database names
The "monetdb status" command truncates long database names such that they fit
on your terminal (or, if you're not running on a terminal, so that they fit
within 80 characters). TKP-web then attempts to use the truncated names, and
fails.

By using the long-format status we can work around this. It's still fragile,
though, as we're relying on parsing unstable command line output.